% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_substack_post}
\alias{get_substack_post}
\title{Get Single Post}
\usage{
get_substack_post(publication_url, slug)
}
\arguments{
\item{publication_url}{Character.}

\item{slug}{Character. Post slug.}
}
\value{
A `data.frame` with one row (or zero rows if missing).
}
\description{
Fetch a single post by its slug. Returns a one‐row `data.frame`, or
zero‐row if not found.
}
