% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{set_substack_key}
\alias{set_substack_key}
\title{Set your Substack API Key}
\usage{
set_substack_key(key)
}
\arguments{
\item{key}{A single string: your Substack API key.}
}
\value{
Invisibly \code{TRUE} if successful.
}
\description{
Store your key once per session. Subsequent calls to other
functions will use this key automatically.
}
